+++ docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for Typopo
 
+## 2.2.3 // 2019-10-01
+### New features
+* Remove extra white spaces between a number sign (#) and a number, e.g # 9 → #9
+* Remove extra white spaces before ordinal indicators (e.g. 1 st → 1st, 1 . spoj → 1. spoj)
+
 ## 2.2.2 // 2019-10-01
 ### Improvements
 * Fix dashes between dates, e.g. 2019 - 08-02 → 2019–08–02 (places ndash)